IMPORTANT UPDATE (see description)
Important update but it does not change the results because it corrects two errors that cancelled each other.

in manual defect initialisation (single or pair), I had exchanged i and j indices.
In the visualisation, I was transposing the matrix.

In the new version, I correctly implement the defect initialisation and there is no transpose operation anymore before plotting.

All the phenomenology remains the same before and after the update, the code is just now correct and cleaner.